% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distro.R
\name{distro}
\alias{distro}
\title{Get info on the system's Linux distribution}
\usage{
distro()
}
\value{
On Linux distributions, it returns a list of four elements:
\itemize{
\item \code{id}, e.g. "ubuntu", "debian"
\item \code{version}, e.g. "18.04", "7.7.1908"
\item \code{codename}, e.g. "bionic", "bullseye"
\item \code{short_version}, which for Ubuntu is the \code{yy.mm} version ("18.04") but for
all other distributions is the major version number only ("7")
}

Not all fields are guaranteed to be defined on all systems. On non-Linux
systems, the function returns \code{NULL}.
}
\description{
Get info on the system's Linux distribution
}
\examples{
distro()
}
